server/requirements.txt
django
mysqlclient
djangorestframework
django-filter
django-cors-headers